requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.4
pytest==8.2.1
gspread>=5.12.0
gspread-dataframe>=3.3.1
google-auth>=2.22.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
lxml>=4.9.3
certifi>=2023.7.22
